Added files via upload
Here is the package of what I have for the registration, log in & log out page. 
You guys can use this and develop from here.If you run into some trouble, please let the team know. 
I believe the source code for the whole project is due this friday the 15th.

-Gia